Redesign interval / schedule models
Why: found the node-schedule JS module. Which uses cron strings for
scheduling jobs.

* ...

Instead of using the DB as a state container, will instead store only
data needed to initialize recuring jobs on application startup.
Jobs should be stored in a hash using databse unique IDs as a key.
This will allow in-memory jobs to be canceled when an entry is removed
from the database.
Also, will store only the topic and message content to be sent.

* ...