Build getWindowedValues helper method
Why: each trigger should know how to get the measurements that
are within the smoothing window.

* ...

Each trigger will be able to fetch measurements based on the
nodemcu and sensorNum it is associated with

the smoothingWindow will be an integer for the number of minutes
into the past to search for measurements.

getWindowedValues will return a promise containing an array of
measurement objects that fall within the smoothing window.

* ...