Implement naive trigger behavior
Why: There is room to optimize how triggers work. This is a
naive approach with simple conditional logic.

* ...

Implemented a trigger model that will check for changes in sensor values
at a pre-defined interval. If the value of the sensor exceeds the
conditions set for the trigger a message will be sent.

Currently: a new message will be sent each time the trigger is is
evaluated, an optimized approach would be to only send a message
when the condition enters or leaves the predefined boundaries.

* ...